+++ config/splat.pspeu.bobo0.yaml
@@ -49,12 +49,11 @@ segments:
       - [0x24A8, c, bo0/e_skelerang]
       - [0x33C8, c, bo0/e_life_up]
       - [0x4048, c, bo0/e_stage_name]
-      - [0x4EC8, c, bo0_psp/e_marionette]
+      - [0x4EC8, c, bo0/e_marionette]
       - [0x6B90, c, bo0/e_red_door]
       - [0x7E98, c, bo0/st_update]
       - [0x83D0, c, bo0/e_particles]
-      - [0x9960, c, bo0_psp/unk_9960]
-      - [0xA2E8, c, bo0/e_collect]
+      - [0x9960, c, bo0/e_collect]
       - [0xDBA8, c, bo0/e_misc]
       - [0xF548, c, bo0/st_collision]
       - [0x11720, c, bo0/popup]
@@ -62,24 +61,25 @@ segments:
       - [0x11E30, c, bo0/e_owl_knight]
       - [0x15A18, c, bo0/st_common]
       - [0x172F8, c, bo0/prim_helpers]
-      - [0x17E18, c, bo0_psp/unk_17E18]
+      - [0x17E18, c, bo0_psp/2B9EC]
       - [0x19A18, c, bo0_psp/st_init_psp]
       - [0x19D60, c, bo0/e_breakable]
-      - [0x19EE0, c, bo0_psp/unk_19EE0]
-      - [0x1B260, c, bo0_psp/e_background_house]
-      - [0x1C208, c, bo0_psp/e_chair]
-      - [0x1CCB0, c, bo0_psp/33D8C]
+      - [0x19EE0, c, bo0_psp/3B014]
+      - [0x1B260, c, bo0/e_background_house]
+      - [0x1C208, c, bo0/e_chair]
+      - [0x1C638, c, bo0/34DA4]
+      - [0x1CCB0, c, bo0/no2_bg]
       - [0x1D4C8, c, bo0/e_room_bg]
-      - [0x1D5A8, c, bo0_psp/unk_1D5A8]
+      - [0x1D5A8, c, bo0_psp/2D26C]
       - [0x31800, c, bo0/e_prisoner]
       - [0x31E80, c, bo0/e_lock_camera]
-      - [0x320E0, c, bo0_psp/unk_320E0]
-      - [0x32790, c, bo0_psp/372E8]
+      - [0x320E0, c, bo0_psp/3AB18]
+      - [0x32790, c, bo0/372E8]
       - [0x32B30, c, bo0/e_spikes]
       - [0x33480, .data, bo0/e_bloody_zombie]
       - [0x33580, .data, bo0/e_skelerang]
       - [0x33670, .data, bo0/e_life_up]
-      - [0x336C0, data]
+      - [0x336C0, .data, bo0/e_marionette]
       - [0x33878, .data, bo0/e_red_door]
       - [0x33890, .data, bo0/st_update]
       - [0x338C0, .data, bo0/e_particles]
@@ -99,22 +99,28 @@ segments:
       - [0x3E960, data, bo0_psp/rooms]
       - [0x3E980, .data, bo0/e_breakable]
       - [0x3E9D0, data]
-      - [0x57E20, data, bo0_psp/unk_1D5A8]
+      - [0x57BE8, .data, bo0/e_background_house]
+      - [0x57D78, .data, bo0/e_chair]
+      - [0x57DB0, .data, bo0/34DA4]
+      - [0x57DB8, .data, bo0/no2_bg]
+      - [0x57E00, data]
+      - [0x57E20, data, bo0_psp/2D26C]
       - [0x58618, .data, bo0/e_prisoner]
       - [0x58690, .data, bo0/e_lock_camera]
-      - [0x58708, data]
+      - [0x58708, data, bo0_psp/3AB18]
+      - [0x58778, .data, bo0/372E8]
       - [0x587F8, .data, bo0/e_spikes]
       - [0x58808, data, __padding]
       - [0x58880, .rodata, bo0/e_bloody_zombie]
       - [0x588A8, .rodata, bo0/e_skelerang]
       - [0x588F0, .rodata, bo0/e_life_up]
-      - [0x58910, .rodata, bo0_psp/e_marionette]
+      - [0x58910, .rodata, bo0/e_marionette]
       - [0x58940, .rodata, bo0/e_red_door]
       - [0x58958, .rodata, bo0/e_collect]
       - [0x589B8, .rodata, bo0/e_owl_knight]
-      - [0x58A18, .rodata, bo0_psp/unk_17E18]
-      - [0x58A30, .rodata, bo0_psp/unk_19EE0]
-      - [0x58A48, .rodata, bo0_psp/unk_1D5A8]
+      - [0x58A18, .rodata, bo0_psp/2B9EC]
+      - [0x58A30, .rodata, bo0_psp/3B014]
+      - [0x58A48, .rodata, bo0_psp/2D26C]
       - {start: 0x58C80, type: .bss, vram: 0x09290300, name: bo0/create_entity}
       - {start: 0x58C80, type: bss, vram: 0x09290320, name: bo0_psp/09290320}
       - {start: 0x58C80, type: .bss, vram: 0x09290328, name: bo0/st_update}
