+++ config/splat.us.bobo0.yaml
@@ -42,11 +42,15 @@ segments:
       - [0x0, data]
       - [0x79C, .data, e_lock_camera]
       - [0x81C, .data, e_breakable]
-      - [0x87C, data]
+      - [0x87C, .data, no2_bg]
+      - [0x8B8, .data, e_background_house]
       - [0xA38, .data, e_spikes]
       - [0xA48, data]
       - [0xB44, .data, e_prisoner]
       - [0xBA8, data]
+      - [0x1300, .data, 372E8]
+      - [0x1374, data, 3AB18]
+      - [0x13D4, data]
       - [0x1458, .data, st_update]
       - [0x14A4, .data, st_collision]
       - [0x1864, .data, e_red_door]
@@ -54,17 +58,19 @@ segments:
       - [0x1A7C, .data, e_collect]
       - [0x1DD0, .data, e_misc]
       - [0x1E88, .data, e_skelerang]
-      - [0x1F74, data, e_marionette]
+      - [0x1F74, .data, e_marionette]
       - [0x2110, .data, e_bloody_zombie]
       - [0x2208, .data, e_owl_knight]
       - [0x24B4, .data, e_life_up]
       - [0x2508, .data, e_particles]
       - [0x2588, .data, e_room_fg]
       - [0x2614, data]
+      - [0x292B0, .rodata, 34DA4]
       - [0x292B8, .rodata, 2B9EC]
       - [0x292F0, .rodata, e_prisoner]
       - [0x29344, .rodata, 2D26C]
       - [0x29678, .rodata, 3AB18]
+      - [0x29684, .rodata, 3B014]
       - [0x2969C, .rodata, e_chair]
       - [0x296E8, .rodata, e_red_door]
       - [0x29700, .rodata, e_collect]
@@ -78,7 +84,7 @@ segments:
       - [0x299FC, c, e_room_bg]
       - [0x29AC4, c, e_lock_camera]
       - [0x29C80, c, e_breakable]
-      - [0x29DB4, c, 33D8C]
+      - [0x29DB4, c, no2_bg]
       - [0x2A3B0, c, e_background_house]
       - [0x2ADCC, c, 34DA4]
       - [0x2B2DC, c, e_spikes]
@@ -87,6 +93,7 @@ segments:
       - [0x2D26C, c]
       - [0x3A880, c, 372E8]
       - [0x3AB18, c]
+      - [0x3B014, c]
       - [0x3BD24, c, e_chair]
       - [0x3C0DC, c, st_update]
       - [0x3C510, c, st_collision]
