+++ config/symbols.pspeu.txt
@@ -218,6 +218,7 @@ g_Status = 0x091CF400; // size:0x334
 g_api = 0x091CF738;
 g_api_FreePrimitives = 0x091CF778;
 g_api_AllocPrimitives = 0x091CF77C;
+g_api_CheckCollision = 0x091CF780;
 g_api_func_80102CD8 = 0x091CF784;
 g_api_PlaySfx = 0x091CF7A0;
 g_api_func_800EDB58 = 0x091CF7A4;
@@ -271,7 +272,7 @@ D_8006C378 = 0x091FC488;
 D_80097924 = 0x091FC490;
 D_8003C730 = 0x091FC498;
 g_pads = 0x091FC4A8; // size:0x10
-g_GpuUsage = 0x091FC4B8;
+g_GpuUsage = 0x091FC4B8; // size:0x24
 g_CurrentBuffer = 0x091FC4E0;
 g_GpuBuffers = 0x091FC4E8; // size:0x38238
 D_8006C384 = 0x09234720;
