--- src/boss/bo0/33D8C.c
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "bo0.h"
-
-INCLUDE_ASM("boss/bo0/nonmatchings/33D8C", func_us_801B3D8C);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/33D8C", func_us_801B3F30);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/33D8C", func_us_801B4148);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/33D8C", func_us_801B41A4);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/33D8C", func_us_801B4210);
