+++ src/boss/bo0/34DA4.c
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "bo0.h"
 
-INCLUDE_ASM("boss/bo0/nonmatchings/34DA4", func_us_801B4DA4);
+#include "../../st/no2_unk_34DA4.h"
