+++ src/boss/bo0/372E8.c
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "bo0.h"
 
-INCLUDE_ASM("boss/bo0/nonmatchings/372E8", func_us_801B72E8);
+#include "../../st/no2_unk_372E8.h"
