+++ src/boss/bo0/3B014.c
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "bo0.h"
+
+INCLUDE_ASM("boss/bo0/nonmatchings/3B014", func_us_801BB014);
+
+INCLUDE_ASM("boss/bo0/nonmatchings/3B014", func_us_801BB08C);
