+++ src/boss/bo0/e_background_house.c
@@ -1,12 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "bo0.h"
 
-INCLUDE_ASM("boss/bo0/nonmatchings/e_background_house", DrawFacade);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/e_background_house", DrawSides);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/e_background_house", DrawRoof);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/e_background_house", Entity3DHouseSpawner);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/e_background_house", Entity3DBackgroundHouse);
+#include "../../st/e_background_house.h"
