+++ src/boss/bo0/e_chair.c
@@ -1,8 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "bo0.h"
 
-INCLUDE_ASM("boss/bo0/nonmatchings/e_chair", func_us_801BBD24);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/e_chair", EntityChair);
-
-INCLUDE_ASM("boss/bo0/nonmatchings/e_chair", func_us_801BBF98);
+#define CHAIR_DEBUG
+#include "../../st/e_chair.h"
