+++ src/boss/bo0/no2_bg.c
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "bo0.h"
 
-INCLUDE_ASM("boss/bo0/nonmatchings/3AB18", func_us_801BAB18);
+#include "../../st/no2_bg.h"
