+++ src/boss/bo0_psp/2B9EC.c
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "bo0.h"
+
+INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/2B9EC", func_pspeu_0924F498);
+
+INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/2B9EC", func_pspeu_0924FDC0);
+
+INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/2B9EC", func_pspeu_09250568);
+
+INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/2B9EC", func_pspeu_092507C0);
+
+INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/2B9EC", func_pspeu_092509C0);
