+++ src/boss/bo0_psp/3AB18.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "bo0.h"
+
+INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/3AB18", func_pspeu_09269760);
