--- src/boss/bo0_psp/e_background_house.c
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "bo0.h"
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_background_house", DrawFacade);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_background_house", DrawSides);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_background_house", DrawRoof);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_background_house", Entity3DHouseSpawner);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_background_house", Entity3DBackgroundHouse);
