--- src/boss/bo0_psp/e_chair.c
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "bo0.h"
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_chair", func_us_801BEDD8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_chair", EntityChair);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_chair", func_pspeu_09253CB8);
