--- src/boss/bo0_psp/e_marionette.c
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "bo0.h"
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_marionette", func_us_801D6254);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_marionette", func_us_801D6474);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/e_marionette", EntityMarionette);
