--- src/boss/bo0_psp/unk_19EE0.c
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "bo0.h"
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_19EE0", func_pspeu_09251560);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_19EE0", func_psp_0923D390);
