--- src/boss/bo0_psp/unk_1D5A8.c
@@ -1,120 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "bo0.h"
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09254C28);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09255118);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09255288);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09256188);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09256308);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09256490);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09256708);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09256A10);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09257F58);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09258530);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09258688);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09258B10);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09258DE0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09258EF0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09258F58);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925B1B0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925C178);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925C580);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925C938);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925CBE8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", EntityOlroxAfterImage);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925D328);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925DBD0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E070);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_psp_09239568);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E1B0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E340);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E488);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E578);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E688);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E778);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925E8C8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925EA30);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925EAD8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925EC00);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925ECA8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925ED50);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925EDF8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925EF08);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925EFF8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925F0A8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925F158);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925F530);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_0925F7A0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_092608C0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_092611E8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09264380);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09264460);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09265170);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_092652C0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09265368);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_092657B8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09265CA0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09265D78);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_092667C8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09267730);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09267E48);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09267EF8);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_1D5A8", func_pspeu_09267FC0);
