--- src/boss/bo0_psp/unk_9960.c
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "bo0.h"
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_9960", func_psp_09240078);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_9960", func_psp_09240110);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_9960", func_psp_09240150);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_9960", func_psp_09240708);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_9960", func_psp_092407E0);
-
-INCLUDE_ASM("boss/bo0_psp/nonmatchings/bo0_psp/unk_9960", func_psp_09240960);
