+++ src/st/blit_char_psp.h
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-u8 func_psp_0923C2F8(u8 arg0) {
+static u8 func_psp_0923C2F8(u8 arg0) {
     if (arg0 & 0x100) {
         return func_psp_0923C2F8(arg0 & 0xFF);
     }
@@ -14,7 +14,7 @@ static u8* pixPtr;
 static s32 frameBufferX;
 static s32 frameBufferY;
 
-void func_psp_0923C390(s32 posX, s32 posY) {
+static void func_psp_0923C390(s32 posX, s32 posY) {
     pixPtr = g_Pix[0];
     frameBufferX = posX;
     frameBufferY = posY;
