+++ src/st/e_collect.h
@@ -201,7 +201,7 @@ static void PrizeDropFall2(u16 arg0) {
 }
 
 // This function is messy, maybe there's a better way.
-void CollectHeart(u16 arg0) {
+static void CollectHeart(u16 arg0) {
 #if defined VERSION_BETA || STAGE == STAGE_ST0
     s8 heartPrizes[2] = {0x01, 0x05};
     s8 mad_unknown[2] = {0x01, 0x02};
