+++ src/st/no2_unk_372E8.h
@@ -1,0 +1,100 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+static u8 D_us_80180EA8[] = {3, 1, 3, 2, 3, 3,  3, 4,  3, 5,  3, 6,  3,  7,
+                             3, 8, 3, 9, 3, 10, 3, 11, 3, 12, 3, 13, -1, 0};
+static u8 D_us_80180EC4[] = {
+    3, 1, 3, 2, 3, 3, 3, 4, 3, 5, 3, 6, 3, 7, 3, 8, -1, 0,
+};
+static u8 D_us_80180ED8[] = {2, 1, 2, 2, 2,  3, 2,  4, 2,  5, 2,  6, 2,  7,  2,
+                             8, 2, 9, 2, 10, 2, 11, 2, 12, 2, 13, 2, 14, -1, 0};
+
+static EntityConfig D_us_80180EF8[] = {
+    {ANIMSET_OVL(10), 0x4D, 0, BLEND_NO, D_us_80180EA8},
+    {ANIMSET_OVL(11), 0x56, 3, BLEND_TRANSP | BLEND_ADD, D_us_80180EC4},
+    {ANIMSET_OVL(12), 0x53, 9, BLEND_NO, D_us_80180ED8},
+};
+
+extern EInit g_EInitParticle;
+
+void func_us_801B72E8(Entity* self) {
+    s32 params;
+    EntityConfig* obj;
+
+    switch (self->step) {
+    case 0:
+        InitializeEntity(g_EInitParticle);
+        params = self->params & 0xF;
+        obj = &D_us_80180EF8[params];
+        self->palette = obj->palette + 0x226;
+        self->blendMode = obj->blendMode;
+        self->animSet = obj->animSet;
+        self->unk5A = obj->unk5A;
+        self->ext.et_801B72E8.animData = obj->animData;
+        self->step = params + 1;
+        if (self->params & 0xFF00) {
+            self->zPriority = (self->params & 0xFF00) >> 8;
+        }
+        if (self->params & 0xF0) {
+            self->palette = PAL_FLAG(PAL_UNK_19F);
+            self->blendMode = BLEND_TRANSP;
+            self->facingLeft = 1;
+        }
+        break;
+
+    case 1:
+        MoveEntity();
+        self->velocityY = FIX(-1.0);
+        if (!AnimateEntity(self->ext.et_801B72E8.animData, self)) {
+            DestroyEntity(self);
+        }
+        break;
+
+    case 2:
+        MoveEntity();
+        self->velocityY = FIX(-1.0);
+        switch (self->step_s) {
+        case 0:
+            if (!AnimateEntity(self->ext.et_801B72E8.animData, self)) {
+                self->drawFlags |= ENTITY_SCALEY | ENTITY_SCALEX;
+                self->drawFlags |= ENTITY_OPACITY;
+                self->scaleX = self->scaleY = 256;
+                self->opacity = 128;
+                self->step_s++;
+            }
+            break;
+
+        case 1:
+            self->scaleX = self->scaleY += 4;
+            self->opacity -= 16;
+            if (!self->opacity) {
+                DestroyEntity(self);
+            }
+            break;
+        }
+        break;
+
+    case 3:
+        MoveEntity();
+        if (!AnimateEntity(self->ext.et_801B72E8.animData, self)) {
+            DestroyEntity(self);
+            return;
+        }
+        switch (self->step_s) {
+        case 0:
+            self->drawFlags = ENTITY_OPACITY;
+            self->opacity = 128;
+            self->step_s++;
+            break;
+
+        case 1:
+            if (self->pose == 5) {
+                self->step_s++;
+            }
+            break;
+
+        case 2:
+            self->opacity -= 4;
+            break;
+        }
+        break;
+    }
+}
